Question 수정/삭제 해당 로그인유저만 할 수 있도록 수정

+++ src/main/java/com/codessquad/qna/question/QuestionController.java
@@ -1,6 +1,6 @@
 package com.codessquad.qna.question;
 
-import com.codessquad.qna.sessionutils.HttpSessionUtils;
+import com.codessquad.qna.utils.HttpSessionUtils;
 import com.codessquad.qna.user.User;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -8,14 +8,13 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
-import org.springframework.web.servlet.ModelAndView;
 
 import javax.servlet.http.HttpSession;
 
 @Controller
 @RequestMapping("/questions")
 public class QuestionController {
-    private static Logger log = LoggerFactory.getLogger(QuestionController.class);
+    private Logger log = LoggerFactory.getLogger(QuestionController.class);
 
     @Autowired
     private QuestionRepository questionRepository;
@@ -43,26 +42,56 @@ public class QuestionController {
 
     @GetMapping("/{id}")
     public String showQuestionContents(@PathVariable Long id, Model model) {
-        model.addAttribute("question", questionRepository.findById(id).orElseThrow(IllegalStateException::new));
+        model.addAttribute("question",
+                questionRepository.findById(id).orElseThrow(IllegalStateException::new));
         return "qna/show";
     }
 
     @GetMapping("/{id}/form")
-    public String updateForm(@PathVariable Long id, Model model) {
-        model.addAttribute("question", questionRepository.findById(id).orElseThrow(IllegalStateException::new));
-        return "/qna/updateForm";
+    public String updateForm(@PathVariable Long id, Model model, HttpSession session) {
+        if (HttpSessionUtils.isNoneExistentUser(session)) {
+            return "redirect:/users/loginForm";
+        }
+
+        User loginUser = HttpSessionUtils.getUserFromSession(session);
+        Question question = questionRepository.findById(id).orElseThrow(IllegalStateException::new);
+        if (question.isNotSameWriter(loginUser)) {
+            return "redirect:/users/loginForm";
+        }
+
+        model.addAttribute("question", question);
+        return "qna/updateForm";
     }
 
     @PutMapping("/{id}")
-    public String updateQna(@PathVariable Long id, String title, String contents) {
+    public String updateQuestion(@PathVariable Long id, String title, String contents, HttpSession session) {
+        if (HttpSessionUtils.isNoneExistentUser(session)) {
+            return "redirect:/users/loginForm";
+        }
+
+        User loginUser = HttpSessionUtils.getUserFromSession(session);
         Question question = questionRepository.findById(id).orElseThrow(IllegalStateException::new);
+        if (question.isNotSameWriter(loginUser)) {
+            return "redirect:/users/loginForm";
+        }
+
         question.update(title, contents);
         questionRepository.save(question);
         return String.format("redirect:/questions/%d",id);
     }
 
     @DeleteMapping("/{id}")
-    public String deleteQna(@PathVariable Long id) {
+    public String deleteQuestion(@PathVariable Long id, HttpSession session) {
+        if (HttpSessionUtils.isNoneExistentUser(session)) {
+            return "redirect:/users/loginForm";
+        }
+
+        User loginUser = HttpSessionUtils.getUserFromSession(session);
+        Question question = questionRepository.findById(id).orElseThrow(IllegalStateException::new);
+        if (question.isNotSameWriter(loginUser)) {
+            return "redirect:/users/loginForm";
+        }
+
         questionRepository.deleteById(id);
         return "redirect:/";
     }
